requirements.txt
# Core Framework Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
langchain==0.0.350
langchain-community==0.0.1
crewai==0.1.0
pydantic==2.5.0

# Blockchain Integration
web3==6.12.0
eth-account==0.9.0
eth-hash[pycryptodome]==0.5.2

# Database Support
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pymongo==4.6.0
redis==5.0.1

# HTTP and API
httpx==0.25.2
aiohttp==3.9.1
requests==2.31.0

# Data Processing
pandas==2.1.4
numpy==1.25.2
python-dateutil==2.8.2
python-docx==1.1.0
PyPDF2==3.0.1

# Configuration and Environment
python-dotenv==1.0.0
pyyaml==6.0.1
toml==0.10.2

# Logging and Monitoring
structlog==23.2.0
prometheus-client==0.19.0

# Security
cryptography==41.0.8
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
httpx-test==0.1.0

# Development Tools
black==23.11.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.6.0

# Jupyter Support (for development and analysis)
jupyter==1.0.0
ipykernel==6.27.1

# Additional Utilities
click==8.1.7
rich==13.7.0
typer==0.9.0